+++ mediawiki/mediawiki.py
@@ -16,7 +16,7 @@ from .mediawikipage import (MediaWikiPage)
 from .utilities import (memoize)
 
 URL = 'https://github.com/barrust/mediawiki'
-VERSION = '0.5.1'
+VERSION = '0.5.2'
 
 
 class MediaWiki(object):